+++ sticky_mitten_avatar/avatars/baby.py
@@ -1,3 +1,4 @@
+from typing import List
 import numpy as np
 from ikpy.chain import Chain
 from ikpy.link import URDFLink, OriginLink
@@ -86,3 +87,30 @@ class Baby(Avatar):
                      translation_vector=[0, 0, -0.0625],
                      orientation=[0, 0, 0],
                      rotation=[0, 0, 0])])
+
+    def _get_default_sticky_mitten_profile(self) -> dict:
+        return {"shoulder_pitch": {"mass": 3, "damper": 555, "force": 150, "angular_drag": 200},
+                "shoulder_yaw": {"mass": 3, "damper": 555, "force": 150, "angular_drag": 200},
+                "shoulder_roll": {"mass": 3, "damper": 555, "force": 150, "angular_drag": 200},
+                "elbow": {"mass": 2, "damper": 555, "force": 150, "angular_drag": 200},
+                "wrist_roll": {"mass": 1.5, "damper": 530, "force": 150, "angular_drag": 200},
+                "wrist_pitch": {"mass": 1.5, "damper": 505, "force": 125, "angular_drag": 200}}
+
+    def _get_start_bend_sticky_mitten_profile(self) -> dict:
+        return {"shoulder_pitch": {"mass": 3, "damper": 75, "force": 450, "angular_drag": 1},
+                "shoulder_yaw": {"mass": 3, "damper": 75, "force": 450, "angular_drag": 1},
+                "shoulder_roll": {"mass": 3, "damper": 75, "force": 450, "angular_drag": 1},
+                "elbow": {"mass": 2, "damper": 75, "force": 450, "angular_drag": 1},
+                "wrist_roll": {"mass": 1.5, "damper": 500, "force": 450, "angular_drag": 1},
+                "wrist_pitch": {"mass": 1.5, "damper": 50, "force": 345, "angular_drag": 1}}
+
+    def _get_reset_arm_sticky_mitten_profile(self) -> dict:
+        return {"shoulder_pitch": {"mass": 3, "damper": 0, "force": 250, "angular_drag": 10},
+                "shoulder_yaw": {"mass": 3, "damper": 0, "force": 250, "angular_drag": 10},
+                "shoulder_roll": {"mass": 3, "damper": 0, "force": 250, "angular_drag": 10},
+                "elbow": {"mass": 2, "damper": 0, "force": 250, "angular_drag": 10},
+                "wrist_roll": {"mass": 1.5, "damper": 0, "force": 250, "angular_drag": 10},
+                "wrist_pitch": {"mass": 1.5, "damper": 0, "force": 225, "angular_drag": 10}}
+
+    def _get_mass(self) -> float:
+        return 80